implement pinia for card state
